[⚙️ chore] 결제 함수 분리  #44
